+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 ### Changed
 - Added flag `allow_duplicate_sequence_identifiers` to GenBank parser
+- GenBank parser now handles multiple isoforms of coding and non-coding genes
 
 ## [0.14.0] 2022-2-18
 ### Added